requirements.txt
# 深度学习核心框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 图像处理与数据增强
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1
scikit-image>=0.21.0

# 科学计算与数据处理
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# 评估指标与可视化
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# 模型导出与优化
onnx>=1.14.0
onnxruntime>=1.16.0

# Web 后端 (Flask)
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0

# 工具
PyYAML>=6.0.1
tensorboard>=2.14.0
